backend/requirements.txt
asgiref==3.11.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
Django==5.2.8
djangorestframework==3.16.1
et_xmlfile==2.0.0
Flask==3.1.2
greenlet==3.2.4
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.5
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pathlib==1.0.1
pluggy==1.6.0
psycopg==3.2.13
Pygments==2.19.2
pypi==2.1
pytest==9.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
SQLAlchemy==2.0.44
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
